# contracts/miniapp/b02-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "黑土循环 B02 移动履约服务",
    "description": "司机、店长、第三空间、签收、库存、遥测、报警与实时事件 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/device/telemetry": {
      "post": {
        "tags": [
          "设备遥测"
        ],
        "summary": "Telemetry",
        "operationId": "telemetry_api_v1_device_telemetry_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryBatch"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Telemetry Api V1 Device Telemetry Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TelemetryBatch": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/TelemetrySample"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Samples"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "vehicle_id",
          "samples"
        ],
        "title": "TelemetryBatch"
      },
      "TelemetrySample": {
        "properties": {
          "sampled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sampled At"
          },
          "temperature_c": {
            "type": "number",
            "maximum": 80.0,
            "minimum": -80.0,
            "title": "Temperature C"
          },
          "humidity_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Humidity Pct"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "sampled_at",
          "temperature_c",
          "humidity_pct",
          "latitude",
          "longitude"
        ],
        "title": "TelemetrySample"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}